Make getting list of internal tables more robust. Related to #241
Getting list of internal tables is still tricky. It should work better for MySql but can be still fragile for other DB engines. Going forward we should not exclude named schemas but restrict query to specific schema only (and this schema needs to be configured by admin).

If getting count of records fails for particular table we return error message instead of exception.